Move dump to extensions file.
